refactor: move functions to new files
